requirements.txt
websockets==15.0.1
chromadb==0.6.3
google-api-python-client==2.187.0
google-auth==2.38.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
google-cloud-storage==2.18.0
sentence-transformers==3.0.0
python-dotenv==1.0.0
